Refactor CV -> AI naming
