+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/BotSegments/CVImageCriteriaEvaluator.cs
@@ -3,7 +3,7 @@ using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Linq;
 using RegressionGames.StateRecorder.BotSegments.Models;
-using RegressionGames.StateRecorder.BotSegments.Models.CVService;
+using RegressionGames.StateRecorder.BotSegments.Models.AIService;
 using RegressionGames.StateRecorder.JsonConverters;
 using UnityEngine;
 
@@ -11,7 +11,7 @@ namespace RegressionGames.StateRecorder.BotSegments
 {
 
     /**
-     * <summary>Evaluates CV Image criteria using CVServiceManager to send/receive HTTP requests to a python server for doing the actual CV evaluations.
+     * <summary>Evaluates CV Image criteria using AIServiceManager to send/receive HTTP requests to a python server for doing the actual CV evaluations.
      * Python 'detects' a source image template in a provided screenshot and then this class evaluates those results against the specified bot segment criteria.</summary>
      */
     public static class CVImageCriteriaEvaluator
@@ -155,7 +155,7 @@ namespace RegressionGames.StateRecorder.BotSegments
                         }
 
                         // do NOT await this, let it run async
-                        _ = CVServiceManager.GetInstance().PostCriteriaImageMatch(
+                        _ = AIServiceManager.GetInstance().PostCriteriaImageMatch(
                             request: new CVImageCriteriaRequest()
                             {
                                 screenshot = new CVImageBinaryData()
@@ -213,7 +213,7 @@ namespace RegressionGames.StateRecorder.BotSegments
                             onFailure:
                             () =>
                             {
-                                RGDebug.LogWarning($"CVImageCriteriaEvaluator - Matched - botSegment: {segmentNumber}, index: {index} - Request - onFailure callback - failure invoking CVService image criteria evaluation");
+                                RGDebug.LogWarning($"CVImageCriteriaEvaluator - Matched - botSegment: {segmentNumber}, index: {index} - Request - onFailure callback - failure invoking AIService image criteria evaluation");
                                 lock (_requestTracker)
                                 {
                                     RGDebug.LogVerbose($"CVImageCriteriaEvaluator - Matched - botSegment: {segmentNumber}, index: {index} - Request - onFailure callback - insideLock");
