+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/BotSegments/Models/AIService/CVImageCriteriaRequest.cs
@@ -3,7 +3,7 @@ using System.Threading;
 using RegressionGames.StateRecorder.JsonConverters;
 using UnityEngine;
 
-namespace RegressionGames.StateRecorder.BotSegments.Models.CVService
+namespace RegressionGames.StateRecorder.BotSegments.Models.AIService
 {
     public class CVImageCriteriaRequest
     {
