+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/BotSegments/Models/AIService/CVImageCriteriaRequest.cs.meta
