+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/BotSegments/Models/AIService/CVImageResultList.cs.meta
