+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/BotSegments/Models/AIService/CVTextResultList.cs.meta
