+++ src/gg.regression.unity.bots/Runtime/Scripts/StateRecorder/BotSegments/Models/CVImageBinaryData.cs
@@ -4,7 +4,7 @@ using Newtonsoft.Json;
 using RegressionGames.StateRecorder.BotSegments.JsonConverters;
 using RegressionGames.StateRecorder.JsonConverters;
 
-namespace RegressionGames.StateRecorder.BotSegments.Models.CVService
+namespace RegressionGames.StateRecorder.BotSegments.Models.AIService
 {
     [JsonConverter(typeof(CVImageBinaryDataJsonConverter))]
     public class CVImageBinaryData
